chore: taskManage optimize
